backend/requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
plotly==5.24.1
seaborn==0.13.2
matplotlib==3.9.2
jupyter==1.1.1
python-dotenv==1.0.1
Pillow==10.4.0
openpyxl==3.1.5
xlrd==2.0.1
gunicorn==22.0.0
whitenoise==6.6.0
python-decouple==3.8
django-storages[azure]==1.14.2
dj-database-url==2.1.0
